separate out common functionality into a shared file
